requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
PyYAML
pytest
pytest-asyncio
aiofiles
openai
anthropic
